Added multiple Features to the Status Bar
Added the hub method and javascript method to display status messages. Changed the player card event listener to listen for the Div, should prevent issues of not being able to play. Changed Penalize and Collectpile hub functions to send to all instead of just caller.

Commented out updateAllCards in MyHub.cs and connection.on("updateMessage") in game.js due to a bug causing the game to immediately end after the first turn. Still unsure of why.